Product schema updated, adding/editing products made almost possible

+++ routes/admin/product.js
@@ -15,42 +15,47 @@ router.get('/', function (req,res) {
 });
 
 router.post('/', function (req, res) {
-    // let properties = req.body.property;
-    // let propertiesArray = [];
-    // for(let i=0;i<properties.length-1;i++) {
-    //     propertiesArray.push(new Property({
-    //         key: properties[i],
-    //         value: properties[++i]
-    //     }));
-    // }
+    let properties = req.body.property;
+    let propertiesArray = [];
+    for(let i=0;i<properties.length-1;i++) {
+        propertiesArray.push(new Property({
+            key: properties[i],
+            value: properties[++i]
+        }));
+    }
+    if (typeof(req.body.colors) == 'string')
+        req.body.colors = [req.body.colors]
     // Category.findOne({name: req.body.category}, function (err, foundCategory) {
         // if(!err) {
             let product = new Product({
-                // name: req.body.name,
-                // description: req.body.description,
-                // remainingNumber: req.body.remainingNumber,
-                // price: req.body.price,
-                // hasDiscount: req.body.hasDiscount,
+                productID: req.body.productID,
+                name: req.body.name,
+                status: req.body.status,
+                price: req.body.price,
+                remainingNumber: req.body.remainingNumber,
+                colors: req.body.colors,
+                images: [],
                 // category: foundCategory._id,
-                // properties: propertiesArray,
-                images: []
+                discount: req.body.discount,
+                description: req.body.description,
+                properties: propertiesArray
             });
 
             // saving images and removing them from temp_images
             if (!fs.existsSync('./public/Images/products'))
-                fs.mkdirSync('./public/Images/product');
+                fs.mkdirSync('./public/Images/products');
             fs.mkdirSync('public/Images/products/' + product._id);
 
             // if it is a single images
             if (typeof(req.body.filepond) == 'string') {
                 image = req.body.filepond;
                 fs.renameSync('./temp_images/' + image, './public/Images/products/' + product._id + '/' + image);
-                // product.images.push('Images/products/' + product._id + '/' + image);
+                product.images.push('Images/products/' + product._id + '/' + image);
                 rimraf.sync('./temp_images/');
             } else { // if it is an array of images
                 for (image of req.body.filepond) {
                     fs.renameSync('./temp_images/' + image, './public/Images/products/' + product._id + '/' + image);
-                    // product.images.push('Images/products/' + product._id + '/' + image);
+                    product.images.push('Images/products/' + product._id + '/' + image);
                 }
                 rimraf.sync('./temp_images/');
             }
@@ -58,7 +63,7 @@ router.post('/', function (req, res) {
             product.save(function (err, savedProduct) {
                 if(!err) {
                     console.log(savedProduct);
-                    res.redirect('/');
+                    res.redirect('/adminDashboard');
                 } else {
                     console.log(err);
                 }
@@ -68,26 +73,31 @@ router.post('/', function (req, res) {
 });
 
 router.put('/:id', function (req, res) {
-    // let properties = req.body.property;
-    // let propertiesArray = [];
-    // for(let i=0;i<properties.length-1;i++) {
-    //     propertiesArray.push(new Property({
-    //         key: properties[i],
-    //         value: properties[++i]
-    //     }));
-    // }
+    let properties = req.body.property;
+    let propertiesArray = [];
+    for(let i=0;i<properties.length-1;i++) {
+        propertiesArray.push(new Property({
+            key: properties[i],
+            value: properties[++i]
+        }));
+    }
+    if (typeof(req.body.colors) == 'string')
+        req.body.colors = [req.body.colors]
     // Category.findOne({name: req.body.category}, function (err, foundCategory) {
     //     if(!err) {
-    //         let newProduct = {
-    //             name: req.body.name,
-    //             description: req.body.description,
-    //             remainingNumber: req.body.remainingNumber,
-    //             price: req.body.price,
-    //             hasDiscount: req.body.hasDiscount,
-    //             category: foundCategory._id,
-    //             properties: propertiesArray,
-    //             images: []
-    //         };
+            let newProduct = {
+                productID: req.body.productID,
+                name: req.body.name,
+                status: req.body.status,
+                price: req.body.price,
+                remainingNumber: req.body.remainingNumber,
+                colors: req.body.colors,
+                images: [],
+                // category: foundCategory._id,
+                discount: req.body.discount,
+                description: req.body.description,
+                properties: propertiesArray
+            };
 
             // saving images and removing them from temp_images            
             // if it is a single images
@@ -104,18 +114,17 @@ router.put('/:id', function (req, res) {
                 rimraf.sync('./temp_images/');
             }
 
-    //         Product.findOneAndUpdate({_id: req.params.id},
-    //             newProduct,
-    //             function (err, product) {
-    //                 if(!err) {
+            Product.findOneAndUpdate({_id: req.params.id},
+                newProduct,
+                function (err, product) {
+                    if(!err) {
                         // deleting old images
                         for (image of product.images) {
                             fs.unlinkSync('./public/' + image);
                         }
                         res.redirect('/');
-    //                     res.redirect('/');
-    //                 }
-    //             })
+                    }
+                })
     //     }
     // })
 });
